feat(server): agent register and heartbeat client
Phase 3 Poppy-Server register and heartbeat client.

+++ src/poppy_agent/robot/adapter.py
@@ -23,7 +23,7 @@ class RobotStatus:
 
     connection_status: str
     operational_status: str
-    battery_percent: float | None
+    battery_percent: int | None
     current_execution_id: str | None
